added boss

+++ Space Rhythm Project/src/spaceRhythm/Game/GameObjects/Block.java
@@ -17,12 +17,12 @@ public class Block extends GameObject {
     @Override
     public void render(Graphics g) {
         g.setColor(Color.MAGENTA);
-        g.fillRect(x, y, 32, 32);
+        g.fillRect((int)x, (int)y, 32, 32);
     }
 
     @Override
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 32, 32);
+        return new Rectangle((int)x, (int)y, 32, 32);
 
     }
