+++ Space Rhythm Project/src/spaceRhythm/Game/GameObjects/GameObject.java
@@ -5,14 +5,14 @@ import spaceRhythm.SpriteSheet.SpriteSheet;
 import java.awt.*;
 
 public abstract class GameObject {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
     protected ObjectID ID;
-    protected int velX = 0;
-    protected int velY = 0;
+    protected float velX = 0;
+    protected float velY = 0;
     protected SpriteSheet ss;
 
-    public GameObject(int x, int y, ObjectID ID, SpriteSheet ss) {
+    public GameObject(float x, float y, ObjectID ID, SpriteSheet ss) {
         this.ID = ID;
         this.x = x;
         this.y = y;
@@ -33,23 +33,23 @@ public abstract class GameObject {
         this.ID = ID;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
@@ -57,7 +57,7 @@ public abstract class GameObject {
         this.velX = velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
